Restructure redux folder
