--- src/Styles/app.css
@@ -1,3 +1,0 @@
-.app--container {
-  padding: 9.33rem 6.25rem;
-}
