--- src/Styles/nav.css
@@ -1,13 +1,0 @@
-nav {
-  width: 100%;
-  height: 5.875rem;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  background-color: white;
-  padding: 0 6.25rem;
-}
-
-.nav--link {
-  margin: 0 2.563rem;
-}
